requirements.txt
altgraph==0.17.4
engineering-notation==0.10.0
packaging==24.0
pefile==2023.2.7
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
Pygments==2.17.2
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.4
pyserial==3.5
pywin32==306
pywin32-ctypes==0.2.2
pyzmq==25.1.2
setuptools==69.5.1
six==1.16.0
tk-tools==0.16.0
tornado==6.4
traitlets==5.14.2
wcwidth==0.2.13
